feat and refact: add comment-field validation in upload-img-form and create new validation module
