+++ js/constants.js
@@ -25,5 +25,7 @@ const COMMENTATOR_NAMES = [
 ];
 
 const MAX_COMMENT_SHOWN_COUNT = 5;
+const MAX_HASHTAGS_COUNT = 5;
+const MAX_TEXT_DESCRIPTION_LETTERS = 140;
 
-export { PHOTO_DESCRIPTION_COUNT, PHOTO_DESCRIPTIONS, TEXT_MESSAGES, COMMENTATOR_NAMES, MAX_COMMENT_SHOWN_COUNT };
+export { PHOTO_DESCRIPTION_COUNT, PHOTO_DESCRIPTIONS, TEXT_MESSAGES, COMMENTATOR_NAMES, MAX_COMMENT_SHOWN_COUNT, MAX_HASHTAGS_COUNT, MAX_TEXT_DESCRIPTION_LETTERS };
