+++ js/dom-elements.js
@@ -29,6 +29,7 @@ const imgUploadModalElement = uploadImgForm.querySelector('.img-upload__overlay'
 const imgUploadModalCloseButton = uploadImgForm.querySelector('.img-upload__cancel');
 const imgUploadSubmit = uploadImgForm.querySelector('.img-upload__submit');
 const textHashtagsInput = uploadImgForm.querySelector('.text__hashtags');
+const textDescriptionField = uploadImgForm.querySelector('.text__description');
 
 export {
   photoTemplate,
@@ -52,5 +53,6 @@ export {
   imgUploadModalElement,
   imgUploadModalCloseButton,
   imgUploadSubmit,
-  textHashtagsInput
+  textHashtagsInput,
+  textDescriptionField
 };
