+++ js/main.js
@@ -1,7 +1,7 @@
 import { createPhotosDescriptions } from './data-generators/photos-descriptions.js';
 import { initializePhotoModal } from './photo-modal.js';
 import { renderThumbnails } from './render-thumbnails.js';
-import { initializePhotoUploadModal } from './photoUploadForm.js';
+import { initializePhotoUploadModal } from './photo-upload-form.js';
 
 const usersPhotos = createPhotosDescriptions();
