+++ js/utils/on-escape-keydown.js
@@ -1,11 +1,11 @@
 import { isEscapeKey } from './keyboard-utils.js';
-import { textHashtagsInput } from '../dom-elements.js';
+import { textHashtagsInput, textDescriptionField } from '../dom-elements.js';
 
 function onEscapeKeydown(closeModalFunction) {
   return function (evt) {
     if (isEscapeKey(evt)) {
       evt.preventDefault();
-      if (document.activeElement === textHashtagsInput) {
+      if (document.activeElement === textHashtagsInput || document.activeElement === textDescriptionField) {
         evt.stopPropagation();
         return;
       }
